requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
chromadb==0.4.3
sentence-transformers==3.3.1
llama-cpp-python==0.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
